+++ frameos/src/drivers/waveshare/ePaper/EPD_7in5_V2.c
@@ -4,7 +4,7 @@
 * | Function    :   Electronic paper driver
 * | Info        :
 *----------------
-* |	This version:   V3.0
+* |	This version:   V1.0
 * | Date        :   2023-12-18
 * | Info        :
 *****************************************************************************
@@ -28,8 +28,167 @@
 # THE SOFTWARE.
 #
 ******************************************************************************/
-#include "EPD_7in5_V2.h"
+#include "EPD_7in5_V2_old.h"
 #include "Debug.h"
+#include <time.h>
+
+UBYTE Voltage_Frame_7IN5_V2[]={
+	0x6, 0x3F, 0x3F, 0x11, 0x24, 0x7, 0x17,
+};
+
+UBYTE LUT_VCOM_7IN5_V2[]={	
+	0x0,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x0,	0xF,	0x1,	0xF,	0x1,	0x2,	
+	0x0,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+};						
+
+UBYTE LUT_WW_7IN5_V2[]={	
+	0x10,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x84,	0xF,	0x1,	0xF,	0x1,	0x2,	
+	0x20,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+};
+
+UBYTE LUT_BW_7IN5_V2[]={	
+	0x10,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x84,	0xF,	0x1,	0xF,	0x1,	0x2,	
+	0x20,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+};
+
+UBYTE LUT_WB_7IN5_V2[]={	
+	0x80,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x84,	0xF,	0x1,	0xF,	0x1,	0x2,	
+	0x40,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+};
+
+UBYTE LUT_BB_7IN5_V2[]={	
+	0x80,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x84,	0xF,	0x1,	0xF,	0x1,	0x2,	
+	0x40,	0xF,	0xF,	0x0,	0x0,	0x1,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+	0x0,	0x0,	0x0,	0x0,	0x0,	0x0,	
+};
+
+const unsigned char						
+Lut_all_fresh[]={0x67,	0xBF,	0x3F,	0x0D,	0x00,	0x1C,
+//VCOM					
+0x00,	0x32,	0x32,	0x00,	0x00,	0x01,
+0x00,	0x0A,	0x0A,	0x00,	0x00,	0x00,
+0x00,	0x28,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//WW					
+0x60,	0x32,	0x32,	0x00,	0x00,	0x01,
+0x60,	0x0A,	0x0A,	0x00,	0x00,	0x00,
+0x80,	0x28,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//BW					
+0x60,	0x32,	0x32,	0x00,	0x00,	0x01,
+0x60,	0x0A,	0x0A,	0x00,	0x00,	0x00,
+0x80,	0x28,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//WB					
+0x90,	0x32,	0x32,	0x00,	0x00,	0x01,
+0x60,	0x0A,	0x0A,	0x00,	0x00,	0x00,
+0x40,	0x28,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//BB					
+0x90,	0x32,	0x32,	0x00,	0x00,	0x01,
+0x60,	0x0A,	0x0A,	0x00,	0x00,	0x00,
+0x40,	0x28,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//Reserved					
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,					
+};					
+
+			
+					
+const unsigned char						
+Lut_partial[]={0x67,	0xBF,	0x3F,	0x0D,	0x00,	0x1C,
+//VCOM					
+0x00,	0x14,	0x02,	0x00,	0x00,	0x01,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//WW					
+0x20,	0x14,	0x02,	0x00,	0x00,	0x01,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//BW					
+0x80,	0x14,	0x02,	0x00,	0x00,	0x01,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//WB					
+0x40,	0x14,	0x02,	0x00,	0x00,	0x01,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//BB					
+0x00,	0x14,	0x02,	0x00,	0x00,	0x01,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+0x00,	0x00,	0x00,	0x00,	0x00,	0x00,
+//Reserved					
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,	0xFF,	0xFF,	0xFF,	0xFF,	0xFF,
+0xFF,					
+};
 
 /******************************************************************************
 function :	Software reset
@@ -92,6 +251,32 @@ static void EPD_WaitUntilIdle(void)
 	DEV_Delay_ms(5);      
     Debug("e-Paper busy release\r\n");
 }
+
+static void EPD_7IN5_V2_LUT(UBYTE* lut_vcom,  UBYTE* lut_ww, UBYTE* lut_bw, UBYTE* lut_wb, UBYTE* lut_bb)
+{
+	UBYTE count;
+
+	EPD_SendCommand(0x20); //VCOM	
+	for(count=0; count<42; count++)
+		EPD_SendData(lut_vcom[count]);
+
+	EPD_SendCommand(0x21); //LUTBW
+	for(count=0; count<42; count++)
+		EPD_SendData(lut_ww[count]);
+
+	EPD_SendCommand(0x22); //LUTBW
+	for(count=0; count<42; count++)
+		EPD_SendData(lut_bw[count]);
+
+	EPD_SendCommand(0x23); //LUTWB
+	for(count=0; count<42; count++)
+		EPD_SendData(lut_wb[count]);
+
+	EPD_SendCommand(0x24); //LUTBB
+	for(count=0; count<42; count++)
+		EPD_SendData(lut_bb[count]);
+}
+
 /******************************************************************************
 function :	Turn On Display
 parameter:
@@ -110,138 +295,171 @@ parameter:
 UBYTE EPD_7IN5_V2_Init(void)
 {
     EPD_Reset();
-    EPD_SendCommand(0x01);			//POWER SETTING
-	EPD_SendData(0x07);
-	EPD_SendData(0x07);    //VGH=20V,VGL=-20V
-	EPD_SendData(0x3f);		//VDH=15V
-	EPD_SendData(0x3f);		//VDL=-15V
 
-	//Enhanced display drive(Add 0x06 command)
-	EPD_SendCommand(0x06);			//Booster Soft Start 
-	EPD_SendData(0x17);
-	EPD_SendData(0x17);   
-	EPD_SendData(0x28);		
-	EPD_SendData(0x17);	
+    // EPD_SendCommand(0x01);			//POWER SETTING
+    // EPD_SendData(0x07);
+    // EPD_SendData(0x07);		//VGH=20V,VGL=-20V
+    // EPD_SendData(0x3f);		//VDH=15V
+    // EPD_SendData(0x3f);		//VDL=-15V
 
-	EPD_SendCommand(0x04); //POWER ON
-	DEV_Delay_ms(100); 
-	EPD_WaitUntilIdle();        //waiting for the electronic paper IC to release the idle signal
-
-	EPD_SendCommand(0X00);			//PANNEL SETTING
-	EPD_SendData(0x1F);   //KW-3f   KWR-2F	BWROTP 0f	BWOTP 1f
-
-	EPD_SendCommand(0x61);        	//tres			
-	EPD_SendData(0x03);		//source 800
-	EPD_SendData(0x20);
-	EPD_SendData(0x01);		//gate 480
-	EPD_SendData(0xE0);  
-
-	EPD_SendCommand(0X15);		
-	EPD_SendData(0x00);		
-
-    /*
-        If the screen appears gray, use the annotated initialization command
-    */
-    EPD_SendCommand(0X50);			
-	EPD_SendData(0x10);
-	EPD_SendData(0x07);
-	// EPD_SendCommand(0X50);			
-	// EPD_SendData(0x10);
-	// EPD_SendData(0x17);
-    // EPD_SendCommand(0X52);			
-	// EPD_SendData(0x03);
-
-	EPD_SendCommand(0X60);			//TCON SETTING
-	EPD_SendData(0x22);
+	EPD_SendCommand(0x01);  // power setting
+	EPD_SendData(0x17);  // 1-0=11: internal power
+	EPD_SendData(*(Voltage_Frame_7IN5_V2+6));  // VGH&VGL
+	EPD_SendData(*(Voltage_Frame_7IN5_V2+1));  // VSH
+	EPD_SendData(*(Voltage_Frame_7IN5_V2+2));  //  VSL
+	EPD_SendData(*(Voltage_Frame_7IN5_V2+3));  //  VSHR
 	
+	EPD_SendCommand(0x82);  // VCOM DC Setting
+	EPD_SendData(*(Voltage_Frame_7IN5_V2+4));  // VCOM
+
+	EPD_SendCommand(0x06);  // Booster Setting
+	EPD_SendData(0x27);
+	EPD_SendData(0x27);
+	EPD_SendData(0x2F);
+	EPD_SendData(0x17);
+	
+	EPD_SendCommand(0x30);   // OSC Setting
+	EPD_SendData(*(Voltage_Frame_7IN5_V2+0));  // 2-0=100: N=4  ; 5-3=111: M=7  ;  3C=50Hz     3A=100HZ
+
+    EPD_SendCommand(0x04); //POWER ON
+    DEV_Delay_ms(100);
+    EPD_WaitUntilIdle();
+
+    EPD_SendCommand(0X00);			//PANNEL SETTING
+    EPD_SendData(0x3F);   //KW-3f   KWR-2F	BWROTP 0f	BWOTP 1f
+
+    EPD_SendCommand(0x61);        	//tres
+    EPD_SendData(0x03);		//source 800
+    EPD_SendData(0x20);
+    EPD_SendData(0x01);		//gate 480
+    EPD_SendData(0xE0);
+
+    EPD_SendCommand(0X15);
+    EPD_SendData(0x00);
+
+    EPD_SendCommand(0X50);			//VCOM AND DATA INTERVAL SETTING
+    EPD_SendData(0x10);
+    EPD_SendData(0x00);
+
+    EPD_SendCommand(0X60);			//TCON SETTING
+    EPD_SendData(0x22);
+
+    EPD_SendCommand(0x65);  // Resolution setting
+    EPD_SendData(0x00);
+    EPD_SendData(0x00);//800*480
+    EPD_SendData(0x00);
+    EPD_SendData(0x00);
+	
+	EPD_7IN5_V2_LUT(LUT_VCOM_7IN5_V2, LUT_WW_7IN5_V2, LUT_BW_7IN5_V2, LUT_WB_7IN5_V2, LUT_BB_7IN5_V2);
+	
+    return 0;
+}
+
+void Epaper_LUT_By_MCU( UBYTE* wavedata)
+{
+    UBYTE count;
+    UBYTE VCEND,BDEND,EVS,XON,PLL;
+
+    VCEND=wavedata[0]&0x08;
+    BDEND=(wavedata[1]&0xC0)>>6;
+    EVS=VCEND|BDEND;
+    PLL=(wavedata[0]&0xF0)>>4;
+    XON=wavedata[2]&0xC0;
+
+    EPD_SendCommand(0x52);	      //EVS
+    EPD_SendData(EVS);
+
+    EPD_SendCommand(0x30);			  //PLL setting 
+    EPD_SendData(PLL);
+
+    EPD_SendCommand(0x01);       //Set VGH VGL VSH VSL VSHR
+    EPD_SendData (0x17);
+    EPD_SendData ((*wavedata++)&0x07);  	//VGH/VGL Voltage Level selection
+    EPD_SendData ((*wavedata++)&0x3F);		//VSH for black
+    EPD_SendData ((*wavedata++)&0x3F);  	//VSL for white
+    EPD_SendData ((*wavedata++)&0x3F);  	//VSHR for red
+
+    EPD_SendCommand(0x2A);	      //LUTOPT
+    EPD_SendData(XON);
+    EPD_SendData(*wavedata++);
+
+    EPD_SendCommand(0x82);       //VCOM_DC setting
+    EPD_SendData (*wavedata++);  //Vcom value
+
+
+    EPD_SendCommand(0x20);
+    for(count=0;count<42;count++)
+        EPD_SendData(*wavedata++);
+
+    EPD_SendCommand(0x21);	
+    for(count=0;count<42;count++)
+        EPD_SendData(*wavedata++);
+
+    EPD_SendCommand(0x22);	
+    for(count=0;count<42;count++)
+        EPD_SendData(*wavedata++);
+
+    EPD_SendCommand(0x23);	
+    for(count=0;count<42;count++)
+        EPD_SendData(*wavedata++);
+
+    EPD_SendCommand(0x24);	
+    for(count=0;count<42;count++)
+        EPD_SendData(*wavedata++);
+ 
+	 
+}
+
+UBYTE EPD_7IN5_V2_Init2(void)
+{
+    EPD_Reset();
+
+    EPD_SendCommand(0x00);  // Panel setting   
+    EPD_SendData(0x3F);
+
+    EPD_SendCommand(0x06);  // Booster Setting
+    EPD_SendData(0x17);  
+    EPD_SendData(0x17);
+    EPD_SendData(0x28);
+    EPD_SendData(0x18);
+
+    EPD_SendCommand(0x50);  // VCOM and DATA interval setting    
+    EPD_SendData(0x22);
+    EPD_SendData(0x07);
+
+    EPD_SendCommand(0x60);  // TCON setting
+    EPD_SendData(0x22);  // S-G G-S
+
+    EPD_SendCommand(0x61);  // Resolution setting
+    EPD_SendData(0x03);//800*480
+    EPD_SendData(0x20);
+    EPD_SendData(0x01);
+    EPD_SendData(0xE0); 
+
+    EPD_SendCommand(0x65);  // Resolution setting
+    EPD_SendData(0x00);
+    EPD_SendData(0x00);
+    EPD_SendData(0x00);
+    EPD_SendData(0x00);
+
+    EPD_SendCommand(0x04); //POWER ON
+    DEV_Delay_ms(100);
+    EPD_WaitUntilIdle();    
+
     return 0;
 }
 
 UBYTE EPD_7IN5_V2_Init_Fast(void)
 {
-    EPD_Reset();
-    EPD_SendCommand(0X00);			//PANNEL SETTING
-    EPD_SendData(0x1F);   //KW-3f   KWR-2F	BWROTP 0f	BWOTP 1f
-
-    /*
-        If the screen appears gray, use the annotated initialization command
-    */
-    EPD_SendCommand(0X50);			
-	EPD_SendData(0x10);
-	EPD_SendData(0x07);
-    // EPD_SendCommand(0X50);		
-	// EPD_SendData(0x10);
-	// EPD_SendData(0x17);
-    // EPD_SendCommand(0X52);		
-	// EPD_SendData(0x03);
-
-    EPD_SendCommand(0x04); //POWER ON
-    DEV_Delay_ms(100); 
-	EPD_WaitUntilIdle();        //waiting for the electronic paper IC to release the idle signal
-
-    //Enhanced display drive(Add 0x06 command)
-    EPD_SendCommand(0x06);			//Booster Soft Start 
-    EPD_SendData (0x27);
-    EPD_SendData (0x27);   
-    EPD_SendData (0x18);		
-    EPD_SendData (0x17);		
-
-    EPD_SendCommand(0xE0);
-    EPD_SendData(0x02);
-    EPD_SendCommand(0xE5);
-    EPD_SendData(0x5A);
-	
+    EPD_7IN5_V2_Init2();
+    Epaper_LUT_By_MCU((UBYTE*)Lut_all_fresh);
     return 0;
 }
 
-UBYTE EPD_7IN5_V2_Init_Part(void)
+UBYTE EPD_7IN5_V2_Init_Partial(void)
 {
-    EPD_Reset();
-
-	EPD_SendCommand(0X00);			//PANNEL SETTING
-	EPD_SendData(0x1F);   //KW-3f   KWR-2F	BWROTP 0f	BWOTP 1f
-	
-	EPD_SendCommand(0x04); //POWER ON
-	DEV_Delay_ms(100); 
-	EPD_WaitUntilIdle();        //waiting for the electronic paper IC to release the idle signal
-	
-	EPD_SendCommand(0xE0);
-	EPD_SendData(0x02);
-	EPD_SendCommand(0xE5);
-	EPD_SendData(0x6E);
-	
-    return 0;
-}
-
-/*
-    The feature will only be available on screens sold after 24/10/23
-*/
-UBYTE EPD_7IN5_V2_Init_4Gray(void)
-{
-    EPD_Reset();
-
-	EPD_SendCommand(0X00);			//PANNEL SETTING
-	EPD_SendData(0x1F);   //KW-3f   KWR-2F	BWROTP 0f	BWOTP 1f
-
-    EPD_SendCommand(0X50);			
-	EPD_SendData(0x10);
-	EPD_SendData(0x07);
-	
-	EPD_SendCommand(0x04); //POWER ON
-	DEV_Delay_ms(100); 
-	EPD_WaitUntilIdle();        //waiting for the electronic paper IC to release the idle signal
-	
-    EPD_SendCommand(0x06);			//Booster Soft Start 
-    EPD_SendData (0x27);
-    EPD_SendData (0x27);   
-    EPD_SendData (0x18);		
-    EPD_SendData (0x17);		
-
-	EPD_SendCommand(0xE0);
-	EPD_SendData(0x02);
-	EPD_SendCommand(0xE5);
-	EPD_SendData(0x5F);
-	
+    EPD_7IN5_V2_Init2();
+    Epaper_LUT_By_MCU((UBYTE*)Lut_partial);
     return 0;
 }
 
@@ -306,7 +524,6 @@ void EPD_7IN5_V2_ClearBlack(void)
     
     EPD_7IN5_V2_TurnOnDisplay();
 }
-
 /******************************************************************************
 function :	Sends the image buffer in RAM to e-Paper and displays
 parameter:
@@ -334,122 +551,42 @@ void EPD_7IN5_V2_Display(UBYTE *blackimage)
     EPD_7IN5_V2_TurnOnDisplay();
 }
 
-void EPD_7IN5_V2_Display_Part(UBYTE *blackimage,UDOUBLE x_start, UDOUBLE y_start, UDOUBLE x_end, UDOUBLE y_end)
+void EPD_7IN5_V2_Display_Partial(UBYTE *blackimage,UDOUBLE x_start, UDOUBLE y_start, UDOUBLE x_end, UDOUBLE y_end)
 {
     UDOUBLE Width, Height;
+    
     Width =((x_end - x_start) % 8 == 0)?((x_end - x_start) / 8 ):((x_end - x_start) / 8 + 1);
     Height = y_end - y_start;
 
-    EPD_SendCommand(0x50);
-	EPD_SendData(0xA9);
-	EPD_SendData(0x07);
-
 	EPD_SendCommand(0x91);		//This command makes the display enter partial mode
 	EPD_SendCommand(0x90);		//resolution setting
-	EPD_SendData (x_start/256);
-	EPD_SendData (x_start%256);   //x-start    
+	EPD_SendData(x_start/256);
+	EPD_SendData(x_start%256);   //x-start    
 
-	EPD_SendData (x_end/256);		
-	EPD_SendData (x_end%256-1);  //x-end	
+	EPD_SendData((x_end-1)/256);		
+	EPD_SendData((x_end-1)%256);  //x-end	
 
-	EPD_SendData (y_start/256);  //
-	EPD_SendData (y_start%256);   //y-start    
+	EPD_SendData(y_start/256);  //
+	EPD_SendData(y_start%256);   //y-start    
 
-	EPD_SendData (y_end/256);		
-	EPD_SendData (y_end%256-1);  //y-end
-	EPD_SendData (0x01);
+	EPD_SendData((y_end-1)/256);		
+	EPD_SendData((y_end-1)%256);  //y-end
+	EPD_SendData(0x01);
+	
+    // for (UDOUBLE j = 0; j < Height; j++) {
+    //     for (UDOUBLE i = 0; i < Width; i++) {
+    //         blackimage[i+j*Width] = ~blackimage[i+j*Width]; 
+    //     }
+    // }
     
     EPD_SendCommand(0x13);
     for (UDOUBLE j = 0; j < Height; j++) {
         EPD_SendData2((UBYTE *)(blackimage+j*Width), Width);
     }
-    EPD_7IN5_V2_TurnOnDisplay();
-}
-
-void EPD_7IN5_V2_Display_4Gray(const UBYTE *Image)
-{
-    UDOUBLE i,j,k;
-    UBYTE temp1,temp2,temp3;
-
-    // old  data
-    EPD_SendCommand(0x10);
-    for(i=0; i<48000; i++) {
-        temp3=0;
-        for(j=0; j<2; j++) {
-            temp1 = Image[i*2+j];
-            for(k=0; k<2; k++) {
-                temp2 = temp1&0xC0;
-                if(temp2 == 0xC0)
-                    temp3 |= 0x00;
-                else if(temp2 == 0x00)
-                    temp3 |= 0x01; 
-                else if(temp2 == 0x80)
-                    temp3 |= 0x01; 
-                else //0x40
-                    temp3 |= 0x00; 
-                temp3 <<= 1;
-
-                temp1 <<= 2;
-                temp2 = temp1&0xC0 ;
-                if(temp2 == 0xC0) 
-                    temp3 |= 0x00;
-                else if(temp2 == 0x00) 
-                    temp3 |= 0x01;
-                else if(temp2 == 0x80)
-                    temp3 |= 0x01; 
-                else    //0x40
-                    temp3 |= 0x00;	
-                if(j!=1 || k!=1)
-                    temp3 <<= 1;
-
-                temp1 <<= 2;
-            }
-
-        }
-        EPD_SendData(temp3);
-        // printf("%x",temp3);
-    }
-
-    EPD_SendCommand(0x13);   //write RAM for black(0)/white (1)
-    for(i=0; i<48000; i++) {             //5808*4  46464
-        temp3=0;
-        for(j=0; j<2; j++) {
-            temp1 = Image[i*2+j];
-            for(k=0; k<2; k++) {
-                temp2 = temp1&0xC0 ;
-                if(temp2 == 0xC0)
-                    temp3 |= 0x00;//white
-                else if(temp2 == 0x00)
-                    temp3 |= 0x01;  //black
-                else if(temp2 == 0x80)
-                    temp3 |= 0x00;  //gray1
-                else //0x40
-                    temp3 |= 0x01; //gray2
-                temp3 <<= 1;
-
-                temp1 <<= 2;
-                temp2 = temp1&0xC0 ;
-                if(temp2 == 0xC0)  //white
-                    temp3 |= 0x00;
-                else if(temp2 == 0x00) //black
-                    temp3 |= 0x01;
-                else if(temp2 == 0x80)
-                    temp3 |= 0x00; //gray1
-                else    //0x40
-                    temp3 |= 0x01;	//gray2
-                if(j!=1 || k!=1)
-                    temp3 <<= 1;
-
-                temp1 <<= 2;
-            }
-        }
-        EPD_SendData(temp3);
-        // printf("%x",temp3);
-    }
 
     EPD_7IN5_V2_TurnOnDisplay();
+    EPD_SendCommand(0x92);
 }
-
 
 /******************************************************************************
 function :	Enter sleep mode
@@ -457,8 +594,6 @@ parameter:
 ******************************************************************************/
 void EPD_7IN5_V2_Sleep(void)
 {
-    EPD_SendCommand(0x50);  	
-    EPD_SendData(0XF7);
     EPD_SendCommand(0X02);  	//power off
     EPD_WaitUntilIdle();
     EPD_SendCommand(0X07);  	//deep sleep
