+++ frameos/src/drivers/waveshare/ePaper/EPD_7in5_V2.nim
@@ -5,7 +5,7 @@
 ##  | Function    :   Electronic paper driver
 ##  | Info        :
 ## ----------------
-##  |	This version:   V3.0
+##  |	This version:   V1.0
 ##  | Date        :   2023-12-18
 ##  | Info        :
 ##  1.Remove:ImageBuff[EPD_HEIGHT * EPD_WIDTH / 8]
@@ -41,14 +41,11 @@ const
 
 proc EPD_7IN5_V2_Init*(): UBYTE {.importc: "EPD_7IN5_V2_Init".}
 proc EPD_7IN5_V2_Init_Fast*(): UBYTE {.importc: "EPD_7IN5_V2_Init_Fast".}
-proc EPD_7IN5_V2_Init_Part*(): UBYTE {.importc: "EPD_7IN5_V2_Init_Part".}
-proc EPD_7IN5_V2_Init_4Gray*(): UBYTE {.importc: "EPD_7IN5_V2_Init_4Gray".}
+proc EPD_7IN5_V2_Init_Partial*(): UBYTE {.importc: "EPD_7IN5_V2_Init_Partial".}
 proc EPD_7IN5_V2_Clear*() {.importc: "EPD_7IN5_V2_Clear".}
 proc EPD_7IN5_V2_ClearBlack*() {.importc: "EPD_7IN5_V2_ClearBlack".}
 proc EPD_7IN5_V2_Display*(blackimage: ptr UBYTE) {.importc: "EPD_7IN5_V2_Display".}
-proc EPD_7IN5_V2_Display_Part*(blackimage: ptr UBYTE; x_start: UDOUBLE;
-                              y_start: UDOUBLE; x_end: UDOUBLE; y_end: UDOUBLE) {.
-    importc: "EPD_7IN5_V2_Display_Part".}
-proc EPD_7IN5_V2_Display_4Gray*(Image: ptr UBYTE) {.
-    importc: "EPD_7IN5_V2_Display_4Gray".}
+proc EPD_7IN5_V2_Display_Partial*(blackimage: ptr UBYTE; x_start: UDOUBLE;
+                                 y_start: UDOUBLE; x_end: UDOUBLE; y_end: UDOUBLE) {.
+    importc: "EPD_7IN5_V2_Display_Partial".}
 proc EPD_7IN5_V2_Sleep*() {.importc: "EPD_7IN5_V2_Sleep".}
